easrc_uci/requirements.txt
joblib
numpy
pandas
PyYAML
scikit-learn
torch
ucimlrepo
